requirements.txt
streamlit
python-dotenv
numexpr
tavily-python
langchain
langchain-core
langchain-community
langchain-openai
pydantic
